+++ classes/EntityListsOptions.php
@@ -17,17 +17,14 @@ class EntityListsOptions {
      *  
      * @return type
      */
-    Public Static function getEnabledEntitiesOnBrowser() {
+    Public Static function getEnabledEntities() {
         $enables_types = [];
         
         $types = get_registered_entity_types();
         foreach ($types as $key => $t) {
             if ($key == 'object') {
-                $sub_arr = $t;
-                
-                foreach ($sub_arr as $sub) {
+                foreach ($t as $sub) {
                     $setting = elgg_get_plugin_setting('entity_lists_' . $sub, self::PLUGIN_ID);
-
                     if ($setting == self::EntityLists_YES) {
                         array_push($enables_types, $sub);
                     }                    
@@ -37,5 +34,33 @@ class EntityListsOptions {
         
         return $enables_types;        
     }
+    
+    /**
+     * Check if user list is enabled
+     * 
+     * @return boolean
+     */
+    Public Static function isUserEnabled() {
+        $setting = elgg_get_plugin_setting('entity_lists_user', self::PLUGIN_ID);
+        if ($setting == self::EntityLists_YES) {
+            return true;       
+        }                 
+        
+        return false;        
+    } 
+    
+    /**
+     * Check if group list is enabled
+     * 
+     * @return boolean
+     */
+    Public Static function isGroupEnabled() {
+        $setting = elgg_get_plugin_setting('entity_lists_group', self::PLUGIN_ID);
+        if ($setting == self::EntityLists_YES) {
+            return true;       
+        }                 
+        
+        return false;        
+    } 
           
 }