+++ views/default/admin/entity_lists/object.php
@@ -12,27 +12,60 @@ if (!elgg_is_active_plugin('datatables_api')) {
     return;
 }
 
+$subtype = get_input('e');
+if (!$subtype) {
+    echo elgg_echo('admin:entity_lists:etype:missing');
+    return;
+}
+
+$title = elgg_echo("item:object:{$subtype}");
+
+$dt_options = [
+    'action' => "entity_lists/objects/{$subtype}",
+    'limit' => elgg_get_config('default_limit'),
+];
+
+$dt_options['headers'] = [ 
+    ['name' => 'id', 'label' => elgg_echo('entity_lists:admin:elgg_objects:table:header:id')],
+    ['name' => 'title', 'label' => elgg_echo('entity_lists:admin:elgg_objects:table:header:title')],
+    ['name' => 'owner', 'label' => elgg_echo('entity_lists:admin:elgg_objects:table:header:owner')],
+    ['name' => 'container', 'label' => elgg_echo('entity_lists:admin:elgg_objects:table:header:container')],
+    ['name' => 'access', 'label' => elgg_echo('entity_lists:admin:elgg_objects:table:header:access')],
+    ['name' => 'created', 'label' => elgg_echo('entity_lists:admin:elgg_objects:table:header:created')],
+    ['name' => 'updated', 'label' => elgg_echo('entity_lists:admin:elgg_objects:table:header:updated')],
+    ['name' => 'actions', 'label' => elgg_echo('entity_lists:admin:elgg_objects:table:header:actions')],
+];    
+
+$content = elgg_view('datatables_api/dtapi_ajax', $dt_options);
+
+echo elgg_format_element('div', ['style' => 'margin: 0 0 10px;'], elgg_view_title($title));
+echo elgg_format_element('div', [], $content);
+
+/*
+
+if (!elgg_is_active_plugin('datatables_api')) {
+    echo elgg_echo('admin:entity_lists:datatable_api:missing');
+    return;
+}
+
 $etype = get_input('e');
 if (!$etype) {
     echo elgg_echo('admin:entity_lists:etype:missing');
     return;
 }
 
-// set title
 $title = elgg_echo("item:object:{$etype}");
 
 $options = array(
     'type' => 'object',
     'subtype' => $etype,
     'limit' => 0,
-    'full_view' => false,
-    'view_toggle_type' => false,
 );
 $entities = elgg_get_entities($options);
 
 if ($entities) {
     $dt_options = [];
-    $dt_options['dt_titles'] = array(
+    $dt_options['dt_titles'] = [ 
         elgg_echo('entity_lists:admin:elgg_objects:table:header:id'),
         elgg_echo('entity_lists:admin:elgg_objects:table:header:title'),
         elgg_echo('entity_lists:admin:elgg_objects:table:header:type'),
@@ -42,12 +75,12 @@ if ($entities) {
         elgg_echo('entity_lists:admin:elgg_objects:table:header:created'),
         elgg_echo('entity_lists:admin:elgg_objects:table:header:updated'),
         elgg_echo('entity_lists:admin:elgg_objects:table:header:actions'),
-    );
+    ];
 
     $dt_data = [];
     foreach ($entities as $e) {
         $dt_data_tmp = [];
-
+        
         $owner = get_entity($e->owner_guid);
         $container = get_entity($e->container_guid);
         
@@ -55,11 +88,12 @@ if ($entities) {
         $dt_data_tmp['guid'] = $e->getGUID();
         $dt_data_tmp['title'] = elgg_view('output/url', array(
             'href' => $e->getURL(),
-            'text' => $e->title,
+            'text' => ($e instanceof \ElggObject?$e->title:$e->name),
             'title' => elgg_echo('entity_lists:admin:elgg_objects:view_entity'),
             'is_trusted' => true,
-        ));
-        $dt_data_tmp['type'] = $e->getSubtype();
+        )); 
+        $dt_data_tmp['type'] = $e instanceof \ElggObject?$e->getSubtype():$type;
+        
         $dt_data_tmp['owner'] = elgg_view('output/url', array(
             'href' => $owner->getURL(),
             'text' => ($owner instanceof \ElggUser?$owner->name:$owner->title),
@@ -72,6 +106,7 @@ if ($entities) {
             'title' => elgg_echo('entity_lists:admin:elgg_objects:view_container'),
             'is_trusted' => true,
         ));
+        
         $dt_data_tmp['access'] = $e->access_id;
         $dt_data_tmp['created'] = date("r", $e->time_created);
         $dt_data_tmp['updated'] = date("r", $e->time_updated);