+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### 2.3.2 (Apr 2018)
+* Added: option to show users and groups
+* Updated: code improvements for using server-side options that DataTables provides through datatable-api plugin
+
 ### 2.3.1 (June 2017)
 * **releases:** Initial release